tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-state-scrolled-home-end-scroll.html

<!DOCTYPE html>
<title>@container: scroll-state(scrolled) Home/End keys press scroll</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#scrolled">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  #scroller {
    width: 200px;
    height: 200px;
    container-type: scroll-state;
    overflow-y: scroll;
  }
  #filler {
    height: 600px;
  }
  #target {
    --none: no;
    --top: no;
    --bottom: no;
    @container scroll-state(scrolled: none) {
      --none: yes;
    }
    @container scroll-state(scrolled: top) {
      --top: yes;
    }
    @container scroll-state(scrolled: bottom) {
      --bottom: yes;
    }
  }
</style>
<div id="scroller">
  <div id="filler">
    <div id="target"></div>
  </div>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    const end = "\uE010";
    scroller.focus();
    await new test_driver.Actions()
      .keyDown(end)
      .keyUp(end)
      .send();
    await new test_driver.Actions()
      .keyDown(end)
      .keyUp(end)
      .send();
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--none"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
  }, "End key scroll is an absolute scroll");

  promise_test(async t => {
    const home = "\uE011";
    scroller.focus();
    await new test_driver.Actions()
      .keyDown(home)
      .keyUp(home)
      .send();
    await new test_driver.Actions()
      .keyDown(home)
      .keyUp(home)
      .send();
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--none"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
  }, "Home key scroll is an absolute scroll");

</script>
